test: extract Grok conversation tests so the repo file stays under 128KiB (#2193)
Static guards rejected AgentConversationRepositoryTest.kt as a new
oversized file after the Grok additions. Move those cases out instead
of raising the baseline.
